refactor&fix: make PR review changes
