+++ rig-core/Cargo.toml
@@ -31,8 +31,8 @@ tokio = { version = "1.34.0", features = ["full"] }
 tracing-subscriber = "0.3.18"
 
 [features]
-rig_derive = ["dep:rig-derive"]
+derive = ["dep:rig-derive"]
 
 [[test]]
 name = "embeddable_macro"
-required-features = ["rig_derive"]
+required-features = ["derive"]
