+++ rig-core/examples/rag.rs
@@ -3,9 +3,9 @@ use std::env;
 use rig::{
     completion::Prompt,
     embeddings::builder::DocumentEmbeddings,
+    embeddings::EmbeddingsBuilder,
     providers::openai::{Client, TEXT_EMBEDDING_ADA_002},
     vector_store::in_memory_store::InMemoryVectorStore,
-    EmbeddingsBuilder,
 };
 
 #[tokio::main]
