+++ rig-core/examples/rag_dynamic_tools.rs
@@ -2,10 +2,10 @@ use anyhow::Result;
 use rig::{
     completion::{Prompt, ToolDefinition},
     embeddings::builder::DocumentEmbeddings,
+    embeddings::EmbeddingsBuilder,
     providers::openai::{Client, TEXT_EMBEDDING_ADA_002},
     tool::{Tool, ToolEmbedding, ToolSet},
     vector_store::in_memory_store::InMemoryVectorStore,
-    EmbeddingsBuilder,
 };
 use serde::{Deserialize, Serialize};
 use serde_json::json;
