+++ rig-core/examples/vector_search_cohere.rs
@@ -2,9 +2,9 @@ use std::env;
 
 use rig::{
     embeddings::builder::DocumentEmbeddings,
+    embeddings::EmbeddingsBuilder,
     providers::cohere::{Client, EMBED_ENGLISH_V3},
     vector_store::{in_memory_store::InMemoryVectorStore, VectorStoreIndex},
-    EmbeddingsBuilder,
 };
 
 #[tokio::main]
