+++ rig-core/src/providers/cohere.rs
@@ -13,9 +13,9 @@ use std::collections::HashMap;
 use crate::{
     agent::AgentBuilder,
     completion::{self, CompletionError},
-    embeddings::{self, embedding::EmbeddingError},
+    embeddings::{self, EmbeddingError, EmbeddingsBuilder},
     extractor::ExtractorBuilder,
-    json_utils, EmbeddingsBuilder,
+    json_utils,
 };
 
 use schemars::JsonSchema;
@@ -183,7 +183,7 @@ pub struct EmbeddingModel {
     ndims: usize,
 }
 
-impl embeddings::embedding::EmbeddingModel for EmbeddingModel {
+impl embeddings::EmbeddingModel for EmbeddingModel {
     const MAX_DOCUMENTS: usize = 96;
 
     fn ndims(&self) -> usize {
@@ -193,7 +193,7 @@ impl embeddings::embedding::EmbeddingModel for EmbeddingModel {
     async fn embed_documents(
         &self,
         documents: Vec<String>,
-    ) -> Result<Vec<embeddings::embedding::Embedding>, EmbeddingError> {
+    ) -> Result<Vec<embeddings::Embedding>, EmbeddingError> {
         let response = self
             .client
             .post("/v1/embed")
@@ -222,7 +222,7 @@ impl embeddings::embedding::EmbeddingModel for EmbeddingModel {
                     .embeddings
                     .into_iter()
                     .zip(documents.into_iter())
-                    .map(|(embedding, document)| embeddings::embedding::Embedding {
+                    .map(|(embedding, document)| embeddings::Embedding {
                         document,
                         vec: embedding,
                     })
