+++ rig-core/src/providers/openai.rs
@@ -11,9 +11,9 @@
 use crate::{
     agent::AgentBuilder,
     completion::{self, CompletionError, CompletionRequest},
-    embeddings::{self, embedding::EmbeddingError},
+    embeddings::{self, EmbeddingError, EmbeddingsBuilder},
     extractor::ExtractorBuilder,
-    json_utils, EmbeddingsBuilder,
+    json_utils,
 };
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
@@ -232,7 +232,7 @@ pub struct EmbeddingModel {
     ndims: usize,
 }
 
-impl embeddings::embedding::EmbeddingModel for EmbeddingModel {
+impl embeddings::EmbeddingModel for EmbeddingModel {
     const MAX_DOCUMENTS: usize = 1024;
 
     fn ndims(&self) -> usize {
@@ -242,7 +242,7 @@ impl embeddings::embedding::EmbeddingModel for EmbeddingModel {
     async fn embed_documents(
         &self,
         documents: Vec<String>,
-    ) -> Result<Vec<embeddings::embedding::Embedding>, EmbeddingError> {
+    ) -> Result<Vec<embeddings::Embedding>, EmbeddingError> {
         let response = self
             .client
             .post("/v1/embeddings")
@@ -268,7 +268,7 @@ impl embeddings::embedding::EmbeddingModel for EmbeddingModel {
                     .data
                     .into_iter()
                     .zip(documents.into_iter())
-                    .map(|(embedding, document)| embeddings::embedding::Embedding {
+                    .map(|(embedding, document)| embeddings::Embedding {
                         document,
                         vec: embedding.embedding,
                     })
