+++ rig-lancedb/examples/vector_search_local_ann.rs
@@ -3,11 +3,10 @@ use std::{env, sync::Arc};
 use arrow_array::RecordBatchIterator;
 use fixture::{as_record_batch, schema};
 use lancedb::index::vector::IvfPqIndexBuilder;
-use rig::vector_store::VectorStoreIndex;
 use rig::{
-    embeddings::embedding::EmbeddingModel,
+    embeddings::{EmbeddingModel, EmbeddingsBuilder},
     providers::openai::{Client, TEXT_EMBEDDING_ADA_002},
-    EmbeddingsBuilder,
+    vector_store::VectorStoreIndex,
 };
 use rig_lancedb::{LanceDbVectorStore, SearchParams};
 use serde::Deserialize;
